1065 refactor
