+++ eo-maven-plugin/src/main/java/org/eolang/maven/PullMojo.java
@@ -47,6 +47,7 @@ import org.eolang.maven.objectionary.OyFallbackSwap;
 import org.eolang.maven.objectionary.OyHome;
 import org.eolang.maven.objectionary.OyRemote;
 import org.eolang.maven.util.Home;
+import org.eolang.maven.util.Online;
 import org.eolang.maven.util.Rel;
 
 /**
@@ -119,7 +120,7 @@ public final class PullMojo extends SafeMojo {
 
     @Override
     public void exec() throws IOException {
-        if (!online()) {
+        if (!new Online().value()) {
             Logger.warn(
                 this, "There is not internet connection. Pull skipped"
             );
