+++ eo-maven-plugin/src/test/java/org/eolang/maven/AssembleMojoTest.java
@@ -29,6 +29,7 @@ import org.cactoos.io.InputOf;
 import org.cactoos.set.SetOf;
 import org.eolang.maven.objectionary.Objectionary;
 import org.eolang.maven.util.Home;
+import org.eolang.maven.util.Online;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
@@ -86,7 +87,7 @@ final class AssembleMojoTest {
                     )
                 )
             ),
-            Matchers.is(SafeMojo.online())
+            Matchers.is(new Online().value())
         );
     }
 
@@ -147,7 +148,7 @@ final class AssembleMojoTest {
                     )
                 )
             ),
-            Matchers.is(SafeMojo.online())
+            Matchers.is(new Online().value())
         );
         MatcherAssert.assertThat(
             new Home(target).exists(
